Account for conversion of display ID's to long
The conversion of mPhysicalDisplayId to a long requires that the data
type is propagated.

+++ services/core/java/com/android/server/display/DisplayDevice.java
@@ -35,7 +35,7 @@ abstract class DisplayDevice {
     private final DisplayAdapter mDisplayAdapter;
     private final IBinder mDisplayToken;
     private final String mUniqueId;
-    private int mPhysicalId = -1;
+    private long mPhysicalId = -1;
 
     // The display device does not manage these properties itself, they are set by
     // the display manager service.  The display device shouldn't really be looking at these.
@@ -53,7 +53,7 @@ abstract class DisplayDevice {
     DisplayDeviceInfo mDebugLastLoggedDeviceInfo;
 
     public DisplayDevice(DisplayAdapter displayAdapter, IBinder displayToken, String uniqueId,
-                         int physicalId) {
+                         long physicalId) {
       this(displayAdapter, displayToken, uniqueId);
       mPhysicalId = physicalId;
     }
@@ -97,7 +97,7 @@ abstract class DisplayDevice {
      *
      * @return The display device id.
      */
-    public final int getPhysicalId() {
+    public final long getPhysicalId() {
         return mPhysicalId;
     }
